Update klipfolio DS script's https use to work with node 11 & 12.
The script to check if Klipfolio data sources are refreshing was
breaking when run in node 11+. Turns out this is as a result
of the http/s module using the URL object to parse
the url and for some reason breaking basic auth with username &
password set in the url string.

Setting the auth detail in the options object and as a result leaving
it to `https.get` to compose the url, fixes it!